Move parsing to AppEngineDeployAction (with tests)
