+++ google-cloud-tools-plugin/testSrc/com/google/cloud/tools/intellij/appengine/cloud/AppEngineDeployActionTest.java
@@ -17,6 +17,7 @@
 package com.google.cloud.tools.intellij.appengine.cloud;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
 import static org.mockito.Mockito.when;
 
 import com.google.common.collect.ImmutableMap;
@@ -25,6 +26,7 @@ import com.google.cloud.tools.intellij.login.CredentialedUser;
 import com.google.cloud.tools.intellij.login.GoogleLoginService;
 import com.google.gdt.eclipse.login.common.GoogleLoginState;
 import com.google.gson.Gson;
+import com.google.gson.JsonParseException;
 
 import com.intellij.remoteServer.runtime.deployment.ServerRuntimeInstance.DeploymentOperationCallback;
 import com.intellij.remoteServer.runtime.log.LoggingHandler;
@@ -87,5 +89,76 @@ public class AppEngineDeployActionTest extends BasePluginTestCase {
     assertEquals(refreshToken, jsonMap.get("refresh_token"));
     assertEquals("authorized_user", jsonMap.get("type"));
     credentialFile.delete();
+  }
+
+  @Test
+  public void testDeployOutputJsonParsingOneVersion() {
+    String jsonOutput =
+        "{\n" +
+        "  \"configs\": [],\n" +
+        "  \"versions\": [\n" +
+        "    {\n" +
+        "      \"id\": \"20160429t112518\",\n" +
+        "      \"last_deployed_time\": null,\n" +
+        "      \"project\": \"some-project\",\n" +
+        "      \"service\": \"default\",\n" +
+        "      \"traffic_split\": null,\n" +
+        "      \"version\": null\n" +
+        "    }\n" +
+        "  ]\n" +
+        "}\n";
+
+    AppEngineDeployAction.DeployOutput deployOutput =
+        AppEngineDeployAction.parseDeployOutput(jsonOutput);
+    assertEquals(deployOutput.getVersion(), "20160429t112518");
+    assertEquals(deployOutput.getService(), "default");
+  }
+
+  @Test
+  public void testDeployOutputJsonParsingTwoVersions() {
+    String jsonOutput =
+        "{\n" +
+        "  \"configs\": [],\n" +
+        "  \"versions\": [\n" +
+        "    {\n" +
+        "      \"id\": \"20160429t112518\",\n" +
+        "      \"last_deployed_time\": null,\n" +
+        "      \"project\": \"some-project\",\n" +
+        "      \"service\": \"default\",\n" +
+        "      \"traffic_split\": null,\n" +
+        "      \"version\": null\n" +
+        "    },\n" +
+        "    {\n" +
+        "      \"id\": \"20160429t112518\",\n" +
+        "      \"last_deployed_time\": null,\n" +
+        "      \"project\": \"some-project\",\n" +
+        "      \"service\": \"default\",\n" +
+        "      \"traffic_split\": null,\n" +
+        "      \"version\": null\n" +
+        "    }\n" +
+        "  ]\n" +
+        "}\n";
+
+    try {
+      AppEngineDeployAction.parseDeployOutput(jsonOutput);
+      fail();
+    } catch (JsonParseException e) {
+      // Success! Should throw a JsonParseException.
+    }
+  }
+
+  @Test
+  public void testDeployOutputJsonParsingOldFormat() {
+    String jsonOutput =
+        "{\n" +
+        "  \"default\": \"https://springboot-maven-project.appspot.com\"\n" +
+        "}\n";
+
+    try {
+      AppEngineDeployAction.parseDeployOutput(jsonOutput);
+      fail();
+    } catch (JsonParseException e) {
+      // Success! Should throw a JsonParseException.
+    }
   }
 }
